# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bruma API
  version: 1.0.0
  description: |
    MVP HTTP API (tenants, domains, templates, dispatch, webhooks, logs). Authenticate with `X-API-Key`
    (admin key or per-tenant key). Webhook endpoint uses HMAC signature headers instead of API keys.

    **Swagger UI:** `GET /docs` when `OPENAPI_DOCS_ENABLED` is true or outside production (see deployment config).
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/webhooks/provider:
    post:
      tags: [Webhooks]
      summary: Provider delivery webhook (HMAC-signed raw body)
      description: |
        Send raw JSON body. Signature: `X-Webhook-Signature` or `X-Bruma-Webhook-Signature` (HMAC-SHA256 hex of body using `WEBHOOK_SECRET`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookPayload" }
      responses:
        "204": { description: Applied }
        "400": { description: Invalid JSON or validation }
        "401": { description: Bad signature }
        "404": { description: Unknown provider_message_id }
        "403": { description: Tenant inactive — no update applied }
        "503": { description: WEBHOOK_SECRET not set }
  /v1/logs:
    get:
      tags: [Logs]
      summary: Delivery event log (tenant scoping enforced)
      description: |
        Tenant keys are scoped to their tenant. Admin must pass `tenant_id` (UUID) query parameter.
      parameters:
        - name: tenant_id
          in: query
          schema: { type: string, format: uuid }
          description: Required when using admin API key
        - name: status
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogPage" }
        "400": { description: Invalid tenant_id or cursor }
        "429":
          description: Rate limited (per API key and per tenant)
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds until the client may retry
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    WebhookPayload:
      type: object
      required: [provider_message_id, event]
      properties:
        provider_message_id: { type: string }
        event:
          type: string
          enum: [delivered, bounced, spam_complaint, failed, sent]
        occurred_at: { type: string }
        recipient: { type: string, format: email }
    LogPage:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              occurred_at: { type: string, format: date-time }
              event_name: { type: string }
              dispatch_id: { type: string, format: uuid }
              tenant_id: { type: string, format: uuid }
              dispatch_status: { type: string }
              recipient: { type: [string, "null"] }
              event_key: { type: [string, "null"] }
              correlation_id: { type: [string, "null"] }
        next_cursor: { type: [string, "null"] }
security:
  - ApiKeyAuth: []